Add tests for current getPath function
